Move LabOrder files to more appropriate location (#410)
* Moved laborder files to etor/orders folder

* Renamed ConvertAndSendLabOrderUsecase => ConvertAndSendDemographicsUsecase

* Moved file to match changes in the main package
