+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/etor/orders/LabOrderConverter.java
@@ -1,4 +1,6 @@
-package gov.hhs.cdc.trustedintermediary.etor.demographics;
+package gov.hhs.cdc.trustedintermediary.etor.orders;
+
+import gov.hhs.cdc.trustedintermediary.etor.demographics.Demographics;
 
 /** Interface for converting a demographics object into a lab order. */
 public interface LabOrderConverter {
