+++ app/src/test/groovy/gov/hhs/cdc/trustedintermediary/etor/orders/UnableToSendLabOrderExceptionTest.groovy
@@ -1,5 +1,6 @@
-package gov.hhs.cdc.trustedintermediary.etor.demographics
+package gov.hhs.cdc.trustedintermediary.etor.orders
 
+import gov.hhs.cdc.trustedintermediary.etor.orders.UnableToSendLabOrderException
 import gov.hhs.cdc.trustedintermediary.wrappers.HttpClientException
 import spock.lang.Specification
